# bmt pkg migrate under swarm