--- a/pkg/bmt/interface.go
+++ b/pkg/bmt/interface.go
@@ -19,15 +19,15 @@ const (
 type Hash interface {
 	hash.Hash
 
-	// SetSpan sets the length prefix of BMT hash.
-	SetSpan(int64)
+	// SetMetaLength sets the length prefix of BMT hash.
+	SetMetaToLength(int64)
 
 	// SetSpanBytes sets the length prefix of BMT hash in byte form.
-	SetSpanBytes([]byte)
+	SetMetaBytes([]byte)
+
+	// Hash
+	Hash([]byte) ([]byte, error)
 
 	// Capacity returns the maximum amount of bytes that will be processed by the implementation.
 	Capacity() int
-
-	// WriteSection writes to a specific section of the data to be hashed.
-	WriteSection(idx int, data []byte)
 }