--- a/pkg/bmtpool/bmtpool.go
+++ b/pkg/bmtpool/bmtpool.go
@@ -7,24 +7,25 @@
 package bmtpool
 
 import (
-	bmtlegacy "github.com/ethersphere/bee/pkg/bmt/legacy"
-	"github.com/ethersphere/bee/pkg/bmt/pool"
+	"github.com/ethersphere/bee/pkg/bmt"
 	"github.com/ethersphere/bee/pkg/swarm"
 )
 
-var instance pool.Pooler
+const Capacity = 32
+
+var instance *bmt.Pool
 
 func init() {
-	instance = pool.New(8, swarm.BmtBranches)
+	instance = bmt.NewPool(bmt.NewConf(swarm.NewHasher, swarm.BmtBranches, Capacity))
 }
 
 // Get a bmt Hasher instance.
 // Instances are reset before being returned to the caller.
-func Get() *bmtlegacy.Hasher {
+func Get() *bmt.Hasher {
 	return instance.Get()
 }
 
 // Put a bmt Hasher back into the pool
-func Put(h *bmtlegacy.Hasher) {
+func Put(h *bmt.Hasher) {
 	instance.Put(h)
 }