--- a/pkg/cac/cac.go
+++ b/pkg/cac/cac.go
@@ -63,11 +63,11 @@ func hasher(data []byte) func([]byte) ([]byte, error) {
 		hasher := bmtpool.Get()
 		defer bmtpool.Put(hasher)
 
-		hasher.SetSpanBytes(span)
+		hasher.SetMetaBytes(span)
 		if _, err := hasher.Write(data); err != nil {
 			return nil, err
 		}
-		return hasher.Sum(nil), nil
+		return hasher.Hash(nil)
 	}
 }