This PR is putting back bmt under swarm. This package implements a crucial component of swarm which in turn is only relevant for swarm, hence its natural place is under bee.

The PR does not only migrate the existing repo in https://github.com/ethersphere/bmt to `pkg/bmt`, but also
- simplify and clean up code
- remove unused/useless function
- reorganise tree and hasher fields
- get rid of locking completely (was originally, but got corrupted on the way).. also tested with `-race`
- get rid of section on hasher, use and reuse buffer on tree