@zelig 2021-04-02
OLD:

```

goos: darwin
goarch: amd64
pkg: github.com/ethersphere/bee/pkg/bmt/legacy
BenchmarkBMT
BenchmarkBMT/SHA3_size_4096
BenchmarkBMT/SHA3_size_4096-16         	  110826	     10365 ns/op	     512 B/op	       3 allocs/op
BenchmarkBMT/Baseline_size_4096
BenchmarkBMT/Baseline_size_4096-16     	   42672	     28132 ns/op	   68625 B/op	     391 allocs/op
BenchmarkBMT/REF_size_4096
BenchmarkBMT/REF_size_4096-16          	   14868	     79594 ns/op	   73152 B/op	     445 allocs/op
BenchmarkBMT/BMT_size_4096
BenchmarkBMT/BMT_size_4096-16          	   25887	     46000 ns/op	   70093 B/op	     450 allocs/op
BenchmarkBMT/SHA3_size_2048
BenchmarkBMT/SHA3_size_2048-16         	  207780	      5558 ns/op	     512 B/op	       3 allocs/op
BenchmarkBMT/Baseline_size_2048
BenchmarkBMT/Baseline_size_2048-16     	   66187	     17890 ns/op	   35856 B/op	     199 allocs/op
BenchmarkBMT/REF_size_2048
BenchmarkBMT/REF_size_2048-16          	   14996	     79708 ns/op	   73152 B/op	     445 allocs/op
BenchmarkBMT/BMT_size_2048
BenchmarkBMT/BMT_size_2048-16          	   38167	     31216 ns/op	   35788 B/op	     229 allocs/op
BenchmarkBMT/SHA3_size_1024
BenchmarkBMT/SHA3_size_1024-16         	  356989	      2932 ns/op	     512 B/op	       3 allocs/op
BenchmarkBMT/Baseline_size_1024
BenchmarkBMT/Baseline_size_1024-16     	   94888	     12043 ns/op	   19472 B/op	     103 allocs/op
BenchmarkBMT/REF_size_1024
BenchmarkBMT/REF_size_1024-16          	   14995	     80365 ns/op	   73152 B/op	     445 allocs/op
BenchmarkBMT/BMT_size_1024
BenchmarkBMT/BMT_size_1024-16          	   51885	     22686 ns/op	   18890 B/op	     120 allocs/op
BenchmarkBMT/SHA3_size_512
BenchmarkBMT/SHA3_size_512-16          	  720728	      1607 ns/op	     512 B/op	       3 allocs/op
BenchmarkBMT/Baseline_size_512
BenchmarkBMT/Baseline_size_512-16      	  127203	      8799 ns/op	   11280 B/op	      55 allocs/op
BenchmarkBMT/REF_size_512
BenchmarkBMT/REF_size_512-16           	   15009	     79165 ns/op	   73152 B/op	     445 allocs/op
BenchmarkBMT/BMT_size_512
BenchmarkBMT/BMT_size_512-16           	   68422	     17003 ns/op	   10698 B/op	      67 allocs/op
BenchmarkBMT/SHA3_size_256
BenchmarkBMT/SHA3_size_256-16          	 1305750	       910 ns/op	     512 B/op	       3 allocs/op
BenchmarkBMT/Baseline_size_256
BenchmarkBMT/Baseline_size_256-16      	  181760	      6308 ns/op	    7184 B/op	      31 allocs/op
BenchmarkBMT/REF_size_256
BenchmarkBMT/REF_size_256-16           	   14961	     79719 ns/op	   73152 B/op	     445 allocs/op
BenchmarkBMT/BMT_size_256
BenchmarkBMT/BMT_size_256-16           	   94028	     11856 ns/op	    6857 B/op	      42 allocs/op
BenchmarkBMT/SHA3_size_128
BenchmarkBMT/SHA3_size_128-16          	 2114281	       561 ns/op	     512 B/op	       3 allocs/op
BenchmarkBMT/Baseline_size_128
BenchmarkBMT/Baseline_size_128-16      	  240338	      4807 ns/op	    5136 B/op	      19 allocs/op
BenchmarkBMT/REF_size_128
BenchmarkBMT/REF_size_128-16           	   14779	     79851 ns/op	   73152 B/op	     445 allocs/op
BenchmarkBMT/BMT_size_128
BenchmarkBMT/BMT_size_128-16           	  126242	      9077 ns/op	    5193 B/op	      31 allocs/op
BenchmarkPool
BenchmarkPool/poolsize_1_size_4096
BenchmarkPool/poolsize_1_size_4096-16  	     243	   4908052 ns/op	 7014276 B/op	   45105 allocs/op
BenchmarkPool/poolsize_8_size_4096
BenchmarkPool/poolsize_8_size_4096-16  	     388	   3470043 ns/op	 7015787 B/op	   45116 allocs/op
BenchmarkPool/poolsize_1_size_2048
BenchmarkPool/poolsize_1_size_2048-16  	     358	   3445971 ns/op	 3583870 B/op	   23005 allocs/op
BenchmarkPool/poolsize_8_size_2048
BenchmarkPool/poolsize_8_size_2048-16  	     733	   1742984 ns/op	 3584601 B/op	   23011 allocs/op
BenchmarkPool/poolsize_1_size_1024
BenchmarkPool/poolsize_1_size_1024-16  	     476	   2486565 ns/op	 1894081 B/op	   12104 allocs/op
BenchmarkPool/poolsize_8_size_1024
BenchmarkPool/poolsize_8_size_1024-16  	    1383	    925249 ns/op	 1894497 B/op	   12107 allocs/op
BenchmarkPool/poolsize_1_size_512
BenchmarkPool/poolsize_1_size_512-16   	     607	   1942656 ns/op	 1074741 B/op	    6803 allocs/op
BenchmarkPool/poolsize_8_size_512
BenchmarkPool/poolsize_8_size_512-16   	    2181	    567483 ns/op	 1074975 B/op	    6804 allocs/op
BenchmarkPool/poolsize_1_size_256
BenchmarkPool/poolsize_1_size_256-16   	     787	   1493094 ns/op	  690706 B/op	    4302 allocs/op
BenchmarkPool/poolsize_8_size_256
BenchmarkPool/poolsize_8_size_256-16   	    2997	    406166 ns/op	  690837 B/op	    4303 allocs/op
BenchmarkPool/poolsize_1_size_128
BenchmarkPool/poolsize_1_size_128-16   	     976	   1174681 ns/op	  524451 B/op	    3204 allocs/op
BenchmarkPool/poolsize_8_size_128
BenchmarkPool/poolsize_8_size_128-16   	    3717	    328042 ns/op	  524346 B/op	    3202 allocs/op
PASS
ok  	github.com/ethersphere/bee/pkg/bmt/legacy	55.028s
```


NEW:

```

 :) go test -v ./... -run none -bench .
goos: darwin
goarch: amd64
pkg: github.com/ethersphere/bee/pkg/bmt
BenchmarkBMT
BenchmarkBMT/SHA3_size_4096
BenchmarkBMT/SHA3_size_4096-16         	  102841	     10726 ns/op	     992 B/op	       5 allocs/op
BenchmarkBMT/Baseline_size_4096
BenchmarkBMT/Baseline_size_4096-16     	   33118	     36761 ns/op	  126003 B/op	     637 allocs/op
BenchmarkBMT/REF_size_4096
BenchmarkBMT/REF_size_4096-16          	   14910	     79523 ns/op	   73152 B/op	     445 allocs/op
BenchmarkBMT/BMT_size_4096
BenchmarkBMT/BMT_size_4096-16          	   23839	     49768 ns/op	   66048 B/op	     386 allocs/op
BenchmarkBMT/SHA3_size_2048
BenchmarkBMT/SHA3_size_2048-16         	  197842	      5816 ns/op	     992 B/op	       5 allocs/op
BenchmarkBMT/Baseline_size_2048
BenchmarkBMT/Baseline_size_2048-16     	   53834	     21781 ns/op	   62513 B/op	     317 allocs/op
BenchmarkBMT/REF_size_2048
BenchmarkBMT/REF_size_2048-16          	   14665	     80115 ns/op	   73152 B/op	     445 allocs/op
BenchmarkBMT/BMT_size_2048
BenchmarkBMT/BMT_size_2048-16          	   35178	     34312 ns/op	   36864 B/op	     215 allocs/op
BenchmarkBMT/SHA3_size_1024
BenchmarkBMT/SHA3_size_1024-16         	  375714	      3108 ns/op	     992 B/op	       5 allocs/op
BenchmarkBMT/Baseline_size_1024
BenchmarkBMT/Baseline_size_1024-16     	   82172	     14215 ns/op	   30768 B/op	     157 allocs/op
BenchmarkBMT/REF_size_1024
BenchmarkBMT/REF_size_1024-16          	   14696	     80702 ns/op	   73152 B/op	     445 allocs/op
BenchmarkBMT/BMT_size_1024
BenchmarkBMT/BMT_size_1024-16          	   46714	     25244 ns/op	   20480 B/op	     119 allocs/op
BenchmarkBMT/SHA3_size_512
BenchmarkBMT/SHA3_size_512-16          	  680788	      1767 ns/op	     992 B/op	       5 allocs/op
BenchmarkBMT/Baseline_size_512
BenchmarkBMT/Baseline_size_512-16      	  117007	     10059 ns/op	   14896 B/op	      77 allocs/op
BenchmarkBMT/REF_size_512
BenchmarkBMT/REF_size_512-16           	   14766	     80693 ns/op	   73152 B/op	     445 allocs/op
BenchmarkBMT/BMT_size_512
BenchmarkBMT/BMT_size_512-16           	   60512	     19450 ns/op	   12288 B/op	      71 allocs/op
BenchmarkBMT/SHA3_size_256
BenchmarkBMT/SHA3_size_256-16          	 1000000	      1084 ns/op	     992 B/op	       5 allocs/op
BenchmarkBMT/Baseline_size_256
BenchmarkBMT/Baseline_size_256-16      	  151408	      7523 ns/op	    6960 B/op	      37 allocs/op
BenchmarkBMT/REF_size_256
BenchmarkBMT/REF_size_256-16           	   14575	     81466 ns/op	   73152 B/op	     445 allocs/op
BenchmarkBMT/BMT_size_256
BenchmarkBMT/BMT_size_256-16           	   78123	     14861 ns/op	    8192 B/op	      47 allocs/op
BenchmarkBMT/SHA3_size_128
BenchmarkBMT/SHA3_size_128-16          	 1642686	       718 ns/op	     992 B/op	       5 allocs/op
BenchmarkBMT/Baseline_size_128
BenchmarkBMT/Baseline_size_128-16      	  208344	      5579 ns/op	    2992 B/op	      17 allocs/op
BenchmarkBMT/REF_size_128
BenchmarkBMT/REF_size_128-16           	   14575	     81770 ns/op	   73152 B/op	     445 allocs/op
BenchmarkBMT/BMT_size_128
BenchmarkBMT/BMT_size_128-16           	   94507	     12403 ns/op	    6144 B/op	      35 allocs/op
BenchmarkPool
BenchmarkPool/poolsize_1_size_4096
BenchmarkPool/poolsize_1_size_4096-16  	     244	   4867025 ns/op	 6624285 B/op	   38910 allocs/op
BenchmarkPool/poolsize_16_size_4096
BenchmarkPool/poolsize_16_size_4096-16 	     340	   3754898 ns/op	 6624724 B/op	   38913 allocs/op
BenchmarkPool/poolsize_1_size_2048
BenchmarkPool/poolsize_1_size_2048-16  	     342	   3508533 ns/op	 3705500 B/op	   21806 allocs/op
BenchmarkPool/poolsize_16_size_2048
BenchmarkPool/poolsize_16_size_2048-16 	     607	   2041333 ns/op	 3705795 B/op	   21809 allocs/op
BenchmarkPool/poolsize_1_size_1024
BenchmarkPool/poolsize_1_size_1024-16  	     446	   2688582 ns/op	 2067042 B/op	   12206 allocs/op
BenchmarkPool/poolsize_16_size_1024
BenchmarkPool/poolsize_16_size_1024-16 	    1077	   1039102 ns/op	 2067278 B/op	   12208 allocs/op
BenchmarkPool/poolsize_1_size_512
BenchmarkPool/poolsize_1_size_512-16   	     574	   2072540 ns/op	 1247668 B/op	    7404 allocs/op
BenchmarkPool/poolsize_16_size_512
BenchmarkPool/poolsize_16_size_512-16  	    1700	    638601 ns/op	 1247763 B/op	    7405 allocs/op
BenchmarkPool/poolsize_1_size_256
BenchmarkPool/poolsize_1_size_256-16   	     670	   1761121 ns/op	  837976 B/op	    5003 allocs/op
BenchmarkPool/poolsize_16_size_256
BenchmarkPool/poolsize_16_size_256-16  	    2629	    439419 ns/op	  838058 B/op	    5004 allocs/op
BenchmarkPool/poolsize_1_size_128
BenchmarkPool/poolsize_1_size_128-16   	     799	   1479237 ns/op	  633160 B/op	    3803 allocs/op
BenchmarkPool/poolsize_16_size_128
BenchmarkPool/poolsize_16_size_128-16  	    3346	    355854 ns/op	  633261 B/op	    3804 allocs/op
PASS
ok  	github.com/ethersphere/bee/pkg/bmt	53.677s
```